# backend/src/openapi/wire.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WIRE API",
    "version": "1.0.0",
    "description": "REST API for wire transfer intent management, approvals, voice challenges, audit bundles, API keys, and webhooks.\n\nThis API provides:\n- Transfer intent lifecycle (create, update, execute)\n- Voice challenge and proof submission\n- Approval decisions and token management\n- Beneficiary management\n- Policy configuration and simulation\n- Event chain and audit bundle generation\n- API keys and webhooks (developer platform)\n\nAll endpoints are prefixed with `/api/wire`.\n\n**Authentication:**\n- UI sessions: `Authorization: Bearer <token>`\n- Integrations: `Authorization: Basic base64(<apiKey>:<apiSecret>)` (recommended) or `X-API-KEY: <apiKey>` (legacy)\n\n**Error Format:** All errors follow `{ error: string, code?: string, details?: object }` format.\n**Idempotency:** Mutating endpoints support `X-Idempotency-Key` (where applicable)."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/wire",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/webhooks": {
      "get": {
        "tags": ["Webhooks"],
        "summary": "List webhooks",
        "description": "Lists webhooks for the authenticated organization. Requires `webhook:view` permission.",
        "operationId": "listWebhooks",
        "security": [
          { "bearerAuth": [] },
          { "basicAuth": [] },
          { "apiKeyHeader": [] }
        ],
        "responses": {
          "200": {
            "description": "List of webhooks",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "type": "object" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Webhooks"],
        "summary": "Create webhook",
        "description": "Creates a webhook and returns `secret` only once. Requires `webhook:create` permission.",
        "operationId": "createWebhook",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "201": { "description": "Webhook created", "content": { "application/json": { "schema": { "type": "object" } } } }
        }
      }
    },
    "/webhooks/{id}": {
      "patch": {
        "tags": ["Webhooks"],
        "summary": "Update webhook",
        "description": "Updates webhook configuration. Requires `webhook:edit` permission.",
        "operationId": "updateWebhook",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": { "description": "Updated", "content": { "application/json": { "schema": { "type": "object" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["Webhooks"],
        "summary": "Delete webhook",
        "description": "Deletes a webhook. Requires `webhook:delete` permission.",
        "operationId": "deleteWebhook",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "type": "object" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/webhooks/{id}/test": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Send test webhook",
        "description": "Queues a test delivery to the webhook target. Requires `webhook:edit` permission.",
        "operationId": "testWebhook",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "202": { "description": "Queued", "content": { "application/json": { "schema": { "type": "object" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/webhooks/{id}/rotate-secret": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Rotate webhook secret",
        "description": "Rotates webhook signing secret and returns the new secret once. Requires `webhook:edit` permission.",
        "operationId": "rotateWebhookSecret",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Rotated", "content": { "application/json": { "schema": { "type": "object" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "UI session auth (Bearer token)."
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Integration auth: base64(<apiKey>:<apiSecret>)"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "Legacy integration auth (key-only). Prefer Basic auth."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "code": {
            "type": "string",
            "description": "Machine-parseable error code (optional)"
          },
          "details": {
            "type": "object",
            "description": "Additional error details (optional)"
          }
        }
      },
...
    }
  }
}